requirements.txt
tqdm==4.46.0
matplotlib==3.4.0
numpy==1.20.2
pycocotools==2.0.2
PyYAML==5.4.1
tensorboard==2.2.2
tensorboardX==2.1
torch==1.8.1
torchvision==0.9.1
webcolors==1.11.1
